스터디 삭제 요청을 개선하라
스터디 삭제 요청에서 계정 주인 검사를 @PreAuthorize로 했지만, 수정과
마찬가지로 테스트에 어려움이 있어서 내부 로직 검사로 개선합니다

+++ app/src/main/java/com/example/bookclub/application/StudyService.java
@@ -113,8 +113,13 @@ public class StudyService {
         return startDate.isBefore(today) || startDate.isEqual(today);
     }
 
-    public StudyResultDto deleteStudy(Account account, Long id) {
+    public StudyResultDto deleteStudy(String email, Long id) {
         Study study = getStudy(id);
+        Account loginAccount = accountService.findUserByEmail(email);
+        if(!study.getEmail().equals(loginAccount.getEmail())) {
+            throw new AccountNotManagerOfStudyException();
+        }
+
         study.deleteAccounts();
         studyRepository.delete(study);
